+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.0.2 ###
+ - Added taxCategories and shippingCategories to productTypes
+
 ### 1.0.1 ###
  - Also sync what emails are triggered for an order status